refactor(core): Piece 2 — extract mechanical appliers from OperationBuilder
Move the request-parameter, response-header, example, and link OA\* construction
out of OperationBuilder into per-family @internal applier classes
(RequestParameterApplier, ResponseHeaderApplier, ExampleApplier, LinkApplier).
OperationBuilder orchestrates them and keeps field precedence. Pure move —
ExamplesTest snapshots stay byte-identical. OperationBuilder drops from 1391 to
1127 lines despite Piece 1 adding provenance code.

Refs #482
